CRUD functionality with JWT authentication and authorization

+++ backend/controller/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from '../models/userModel.js';
 
 // @desc    Get Goals
 // @routes  GET /api/goals
 // @access  Private
 export const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id}); // retrieve all documents from the goals collection
 
     res
     .status(200)
@@ -24,7 +25,8 @@ export const setGoals = asyncHandler(async (req,res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res
     .status(200)
@@ -41,6 +43,20 @@ export const updateGoals = asyncHandler(async (req,res) => {
         res.status(400);
         throw new Error('Goal not found');
     };
+
+    const user = await User.findById(req.user.id);
+
+    // If user doesn't exist
+    if (!user) {
+        res.status(401);
+        throw new Error('user not found');
+    };
+
+    // CHeck to ensure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true });
 
@@ -59,6 +75,20 @@ export const deleteGoals = asyncHandler(async (req,res) => {
         res.status(400);
         throw new Error('Goal not found');
     };
+
+    const user = await User.findById(req.user.id);
+
+    // If user doesn't exist
+    if (!user) {
+        res.status(401);
+        throw new Error('user not found');
+    };
+
+    // CHeck to ensure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
